requirements.txt
affine
annotated-types
anyio
attrs
cachetools
charset-normalizer
click
click-plugins
cligj
contourpy
cycler
earthengine-api
fastapi
filelock
fonttools
fsspec
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
h11
httplib2
idna
jinja2
kiwisolver
markupsafe
matplotlib
mpmath
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
packaging
patchify
pillow
proto-plus
protobuf
pyasn1
pyasn1-modules
pydantic
pydantic-core
pyparsing
pyproj
python-dateutil
python-multipart
rasterio
requests
rsa
shapely
sniffio
starlette
sympy
torch
torchvision
tqdm
triton
typing-extensions
uritemplate
urllib3
uvicorn
